URI encode the public key in heartbeats
This commit updates the bytes that we send to the Reflect API heartbeat endpoint.
Since the wireguard public key can contain URI-reserved characters,
we need to percent-encode it to replace those reserved characters.
This commit adds the encoding to the public key before using it in the HTTP request.
The Reflect API server will expect the public key to use this percent-encoding.

**Tested**
- Ran the agent locally and verified that it checks-in successfully with the Reflet API