Add support for NAT traversal
Fixes #6

This commit updates the agent to support a new flag, -n, for agents behind NATs.
When set, this flag causes the agent to use wireguard's keepalive heartbeat.
Additionally, this sends a raw UDP datagram to ensure the hole is punched.

In relation to this change, the Reflect API was updated to include information
about active browser sessions, which allows the agent to initiate each connection.
Previously, the agent acted as the server in the wireguard connection.
This commit reverses the connection so that the agent acts as a client,
which connects to each individual session as a distinct peer in the wireguard network.

Finally, it updates the agent to use more formal argument parsing, and
updates all of the log lines to be more consistent and simpler.

**Tested**
- Ran the agent behind a NAT and verified it registers and checks in
- Verified that new recording and test run sessions are identified and
installed as peers into the wireguard configuration.
- Verified that the sessions in Reflect use the proxy successfully.
- Verified that running the agent without the -n flag is successful