Handle agent reconnection gracefully
The agent's websocket to Reflect might get disconnected at times.
When this happens, the agent reconnects to the endpoint, but
during that time, the messages.txt log that we tail gets reset.
When this happens, the monitor script will see an empty string
and interpret it as the list of active sessions.
When it tries to toggle the wireguard interface, it'll fail
because the empty string appears to be a missing argument.
This, in turn, causes the monitor to exit, and the agent
will never reconfigure its wireguard interface again.

This commit updates the monitor logic to not try to interpret
the latest message as the list of active sessions unless its non-empty.

**Tested**
- Repro'd the bug with a forced reconnection and verified the fix handles it